cmd_bmp: fix very long uncompressing time of gzipped bitmaps
If a compressed bitmap is located in sectors at the end of the
flash and it's offset + CONFIG_SYS_VIDEO_LOGO_MAX_SIZE > 0xFFFFFFFF,
the uncompressing time is very long, since processing the
stream is done bytewise (and not blockwise) due to overflow
in inflate_fast() while calculation and checking for enough
input available.

Fix available bitmap data input size for gunzip() to match
the actually available data size to prevent the observed
misbehaviour.

Reported-by: Werner Pfister <werner.pfister@intercontrol.de>
Signed-off-by: Anatolij Gustschin <agust@denx.de>