Fix bug of get_addr() in fdt_decode.c.
The original description is to match the field holding "one address with
no trailing data". So the if condition should be equal length:
  if (cell && len != sizeof(addr_t))

But it still worked fine except the lcd frame-buffer value because all
other fields are somethings like: reg = <0x70006040 0x40>, not exactly one
address.

So this CL fixed the matching to either one address or two addresses.

BUG=chrome-os-partner:5338
TEST=build without error and manually print the frame-buffer value to verify.

Change-Id: I7a1a64feafc4e883f23044c1d04da09b2122ff40
Reviewed-on: http://gerrit.chromium.org/gerrit/5875
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5973
Reviewed-by: Hung-Te Lin <hungte@chromium.org>