CHERRY-PICK: Fix bug of get_addr() in fdt_decode.c.
The original description is to match the field holding "one address with
no trailing data". So the if condition should be equal length:
  if (cell && len != sizeof(addr_t))

But it still worked fine except the lcd frame-buffer value because all
other fields are somethings like: reg = <0x70006040 0x40>, not exactly one
address.

So this CL fixed the matching to either one address or two addresses.

BUG=chrome-os-partner:5338
TEST=build without error and manually print the frame-buffer value to verify.

Change-Id: Ic5c80fdcb2cb5eed5d0ce20ad91d64a20fa77a7e
Reviewed-on: http://gerrit.chromium.org/gerrit/5875
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6151
Tested-by: Bryan Freed <bfreed@chromium.org>