CHERRY-PICK: Expand flash map to maximize size for U-Boot
This increases the size of RO U-Boot to 1MiB and RW U-Boot to 950KiB
to maximize flexibility in the future. We cannot change the FMAP later
so this is trying to cover for all eventualities.

Note: This commit does not affect Seaboard, since it only has 2MiB
of SPI flash.

BUG=chromium-os:19238
TEST=build for Aebl, Kaen, Seaboard

Change-Id: I1bba03195df3bfad64d683375fce5278e1f4bf88
Original-Change-Id: Ifb57fec75eba06eb9946e6a8d7e8100bd5e92c57
Original-Reviewed-on: http://gerrit.chromium.org/gerrit/6227
Original-Reviewed-by: Randall Spangler <rspangler@chromium.org>
Original-Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Original-Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6352
Reviewed-by: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>