u-boot hack to disable recovery on TPM init failure
This is a hack to disable going to recovery mode in vboot when there is a
TPM failure to initialize during the firmware selection process. We
cold_reboot() the processor when a TPM communication error is detected
instead of going into recovery mode.

Now covering i2c probe calls in ifx_v05.c, as a sanity test a disconnected
TPM now reboots continuously.

BUG=chrome-os-partner:5442
TEST=None

Change-Id: I7e1799bc92707c25c2202891d783e87a19e2b06c
Reviewed-on: http://gerrit.chromium.org/gerrit/9757
Reviewed-by: Micah Catlin <micahc@google.com>
Tested-by: Micah Catlin <micahc@google.com>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>