Modifications to hold the EC in reset during U-Boot
Cherry-pick from 0.15.877.B

This modification was done to avoid TPM communication problems during U-Boot
which can result in a false recovery state.

The reset assertion is added conditionally with a new entry in the FDT passed
to U-Boot.

The GPIO assertion code is semi modular and can be expanded to assert other
GPIO pins.

BUG=chrome-os-partner:5442
TEST=Manual

Change-Id: I73011bbf391430e5d25e1a657b4750f20239b949
Reviewed-on: https://gerrit.chromium.org/gerrit/11012
Reviewed-by: Micah Catlin <micahc@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>