Modifications to hold the EC in reset during U-Boot
This modification was done to avoid TPM communication problems during U-Boot
which can result in a false recovery state.

The reset assertion is added conditionally with a new entry in the FDT passed
to U-Boot.

The GPIO assertion code is semi modular and can be expanded to assert other
GPIO pins.

BUG=chrome-os-partner:5442
TEST=Manual

Change-Id: I73011bbf391430e5d25e1a657b4750f20239b949
Reviewed-on: https://gerrit.chromium.org/gerrit/10536
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>