Add generic relocation feature
Add a relocation implementation as the first thing in the generic board
library. This library is needed by SPL also.

We create a separate header file for link symbols defined by the link
scripts. It is helpful to have these all in one place and try to
make them common across architectures. Since Linux already has a similar
file, we bring this in even though many of the symbols there are not
relevant to us.

The __relocate_code() function is what we expect all architectures which
support relocation will use eventually. For now, they all override this
with their own version.

Note: The conflict with the generic board series is that the file
include/asm-generic/sections.h is created in both. This should be easy
to resolve once we know the order in which these series will be applied.

Series-changes: 2
- Import asm-generic/sections.h from Linux and add U-Boot extras
- Squash generic link symbols patch into generic relocation patch
- Move reloc.c into common/
- Add function comments
- Use memset, memcpy instead of inline code
- Add README file for relocation

Series-changes: 4
- Use renamed start_call_board_init_r() function

Series-changes: 5
- Rename start_call_board_init_r() to pivot_to_board_init_r()

Change-Id: I7963c3577e50d20e0190d2b6d8eb5c00ab9c9e74