arm: Remove unused code in start.S
Now that we are using the generic relocation framework, we don't need this
code.

Note: Here we lose the ARM1176's enable_mmu code. This seems to duplicate
code already in U-Boot now. Can anyone comment on this?

Series-to: u-boot
Series-cc: graeme, mikef, albert, wolfgang
Series-cc: Tom Rini <trini@ti.com>, Jason Liu <r64343@freescale.com>
Series-version: 5
Series-changes: 2
- Make relocation symbols global so we can use them outside start.S

Series-changes: 3
- Remove the 'reloc' tag from each commit
- Rebase to master

Series-changes: 4
- Update start.S pruning to fit with early patches
- Rebase to master, also bring in ARM master

Cover-letter:
Introduce generic relocation feature

(This series has trivial conflicts with the generic board series in the
arch/xxx/config.mk additions and also in that they both add
include/linux/sections.h. I original had this series first in my list,
but the generic board series now also applies directly to master, so I
will leave this one as is until I know what ordering is required.)

This is the second patch series aiming to unify the various board.c files
in each architecture into a single one. This series implements a generic
relocation feature, which is the bridge between board_init_f() and
board_init_r(). It then moves ARM over to use this framework, as an
example.

On ARM the relocation code is duplicated for each CPU yet it
is the same. We can bring this up to the arch level. But since (I believe)
Elf relocation is basically the same process for all archs, there is no
reason not to bring it up to the generic level.

Each architecture which uses this framework needs to provide a function
called arch_elf_relocate_entry() which processes a single relocation
entry. This is a static inline function to reduce code size overhead.

For ARM, a new arch/arm/lib/proc.S file is created, which holds generic
ARM assembler code (things that cannot be written in C and are common
functions used by all ARM CPUs). This helps reduce duplication. Interrupt
handling code and perhaps even some startup code can move there later.

It may be useful for other architectures with a lot of different CPUs
to have a similar file.

Code size on my ARMv7 system increases by 54 bytes with generic
relocation. This overhead is mostly just literal pool access and setting
up to call the relocated U-Boot at the end.

On my system, execution time increases from 10.8ms to 15.6ms due to the
less efficient C implementations of the copy and zero loops. If execution
time is of concern, you can define CONFIG_USE_ARCH_MEMSET and
CONFIG_USE_ARCH_MEMCPY to reduce it. For me this reduces relocation time
to 5.4ms, i.e. twice as fast as the old system.

To address the build failure in tx25 I have split out the memset()/memcpy()
functions into their own file. I believe this is somewhat justified by the
fact that they are now used for relocation, independently of what other
part of U-Boot may or may not use the string functions. The problem with
linking with the whole string.o object file is that the code size is too
large for highly-constrained SPL builds which don't have
-ffunction-sections defined.
END

Change-Id: Iab85ee1779980682ff6923f6710727373c2d6f13