Define CONFIG_SYS_SKIP_RELOC for all archs
We are introducing a new generic relocation features and we want this to
be the default. So we need to opt all architectures out first. Some may
never have relocation, but those that do will eventually move over to
this generic relocation framework.

This is part of the unified board effort, but since we are only dealing
with relocation in this series, CONFIG_SYS_SKIP_RELOC is more appropriate
than CONFIG_SYS_LEGACY_BOARD.

Series-changes: 2
- Use CONFIG_SYS_SKIP_RELOC instead of CONFIG_SYS_LEGACY_BOARD

Series-changes: 4
- Capital D on Define of CONFIG_SYS_SKIP_RELOC commit message

Change-Id: Ice8dd142d6da46633647f6484144edf8e18757e3