Don't use bogus gbb address when  booting up rewriteable firmware.
[cherry pick of dd884e4 previously submitted into cros/chromeos-v2011.06]

https://gerrit.chromium.org/gerrit/11152 introduced a bug where the
gbb pointer used during rewriteable firmware boot is bogus.

The pointer is retrieved from the 'chromeos-config' section of the
device tree, but on x86 platform this memory area is never
initialized. The fix is to make sure that the proper gbb address in
the CPU address space is used before gbb contents are accessed.

What it boils down to is that when CONFIG_HARDWARE_MAPPED_SPI is set,
gbb address should be determined before setup_gbb_and_cdata() is
called.

To accomplish that fdt_decode_twostop_fmap() is being modified to
retrieve the flash base address among other things. Calling this
function before setup_gbb_and_cdata() allows to assign the gbb pointer
before it is used.

`google-binary-block' is not yet being removed from the cromeos-config
section of the device tree as this could break some tests.

BUG=chromium-os:22528
BUG=chrome-os-partner:7155
TEST=manual
  . build a new firmware image and program it on a Lumpy
  . verify that the device comes up as expected
  . modify the firmware to use read/write path as suggested by hungte@
     cd ~/trunk/src/platform/vboot_reference/scripts/image_signing
      ./resign_firmwarefd.sh /build/lumpy/firmware/image.bin \
          /build/lumpy/firmware/new_image.bin \
         ../../tests/devkeys/firmware_data_key.vbprivk \
         ../../tests/devkeys/firmware.keyblock \
         ../../tests/devkeys/firmware_data_key.vbprivk \
         ../../tests/devkeys/firmware.keyblock \
         ../../tests/devkeys/kernel_subkey.vbpubk \
         1 0
  . put new_image.bin into a lumpy flashrom
  . reboot the device
  . observe it come up, printing on the console
    'vboot_twostop: jump to readwrite main firmware at 0x1110000, size 0xdffc0'
    along the way

Change-Id: Ib228157d3f09ce77ca08864aeefb46cf70635b2d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/12969
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/12997