usb-ehci: Support interrupt transfers via periodic list
Interrupt transfers aren't meant to be used from the async list
(the EHCI spec indicates trouble with low/full-speed intr on async).

Build a periodic list instead, and provide an API to make use of it.
Then, use that API from the existing interrupt transfer API.

BUG=chrome-os-partner:5752
TEST=Use USB keyboard in u-boot (recovery mode) when later
commits make use of this code

Change-Id: Ib8276c141e6e3ea0c8829d993bf197614fa9de22
Signed-off-by: Patrick Georgi <patrick@georgi-clan.de>
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/13739
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>