temp: Add CONFIG_SYS_SYM_OFFSETS to support offset symbols
Link symbols as created by the link script can either be absolute or
relative to the text start. This option switches between the two options
so that we can support both.

As we convert architectures over to generic board, we can see if this
option is actually needed, or whether it is possible to unify this feature
also.

Series-changes: 5
- Add offsets support for access to fdt that uses CONFIG_OF_SEPARATE

Change-Id: Id2611805c76c6aa54ca0f1185910739295a06ad7