ppc: Change memsz variable to signed char
This seems to be unsigned char for no good reason. Tidy this up and
remove the casts.

Series-to: u-boot
Series-cc: wolfgang
Series-version: 3
Cover-letter:
Fix a few warnings that bug me
This series fixes some warnings which seem to still be left over from
all the warning-squashing efforts for 2011.12. If patches exist for some
of these already then please ignore.

The alias problem in USB is a bit odd, since I thought it was already
fixed. But perhaps it has come back.
END

Acked-by: Mike Frysinger <vapier@gentoo.org>
Change-Id: Iab1c78d8c62de7cc454ce11d091555d3f6f071f4