tegra: Enable generic board for Seaboard.
Enable CONFIG_SYS_GENERIC_BOARD for Tegra 2 Seaboard.

Series-to: u-boot
Series-cc: wolfgang, albert, graeme, mikef, tomr
Series-cc: vadim, Hung-ying Tyan <tyanh@chromium.org>
Series-version: 5
Series-changes: 2
- Change generic board to an opt-in system on a per-board basic
- Add CONFIG_SYS_GENERIC_BOARD to allow board to select generic board
- Add PowerPC support
- Rebase to master

Series-changes: 3
- Rebase to master
- Rebase on top of x86/master (which has not yet been pulled to master)

Series-changes: 4
- Rebase to master
- Drop sc520_timer.c patch (warning already fixed by previous patch)

Cover-letter:
Create generic board init for ARM, x86, PPC
This series creates a generic board init implementation which contains
the essential functions of the major arch/xxx/lib/board.c files. It is
split into two parts: board_f.c for pre-relocation and board_r.c for
post-relocation.

What is the motivation for this change?

1. There is a lot of repeated code in the board.c files. Any change to
things like setting up the baud rate requires a change in 10 separate
places.

2. Since there are 14 separate files, adding a new feature which requires
initialisation is painful since it must be independently added in 14
places.

3. As time goes by the architectures naturely diverge since there is limited
pressure to compare features or even CONFIG options against simiilar things
in other board.c files.

4. New architectures must implement all the features all over again, and
sometimes in subtley different ways. This places an unfair burden on getting
a new architecture fully functional and running with U-Boot.

5. While it is a bit of a tricky change, I believe it is worthwhile and
achievable. There is no requirement that all code be common, only that
the code that is common should be located in common/board.c rather than
arch/xxx/lib/board.c.

All the functions of board_init_f() and board_init_r() are broken into
separate function calls so that they can easily be included or excluded
for a particular architecture. It also makes it easier to adopt Graeme's
initcall proposal assuming this comes about.

http://lists.denx.de/pipermail/u-boot/2012-January/114499.html

This series is not dependent on generic relocation. So relocation
happens as one big chunk and is still completely arch-specific. See the
relocation series for a proposed solution to this for ARM:

http://lists.denx.de/pipermail/u-boot/2011-December/112928.html

or x86's implementation which is in mainline. Unifying relocation is
probably the next step after this series.

Instead of moving over a whole architecture, this series takes the approach
of simply enabling generic board support for an architecture. It is then up
to each board to opt in by defining CONFIG_SYS_GENERIC_BOARD in the board
config file. If this is not done, then the code will be generated as
before. This allows both sets of code to co-exist until we are comfortable
with the generic approach, and enough boards run.

ARM is a relatively large board.c file and one which I can test, therefore
I think it is a good target for this series. On the other hand, x86 is
relatively small and simple, but different enough that it introduces a
few issues to be solved. So I have chosen both ARM and x86 for this series.
After a suggestion from Wolfgang I have added PPC also. This is the
largest and most feature-full board, so hopefully we have all bases
covered in this series. Other archs are mostly a subset of these.

A generic global_data structure is now in mainline, and this is required
for this series.

Similarly we need a generic bd_info structure, since generic code will
be accessing it. I have done this with a simple generic file for now.

There was dicussion on the list about passing gd_t around as a parameter
to pre-relocation init functions. I think this makes sense, but it can
be done as a separate change, and this series does not require it.

While this series needs to stand on its own, the goal is the unification
of the board init code. So I hope we can address issues with this in mind,
rather than focusing too narrowly on particular ARM, x86 or PPC issues.
I have added TODO markers in places where I think there are opportunities
to relationalise the board init now it is all in one place, but these don't
need to be addressed for the feature to work, and are best done as smaller
patches than can be reviewed by individual arch maintainers, I think.

I have run-tested ARM on Tegra Seaboard and x86/coreboot only. To try it
out, define CONFIG_SYS_GENERIC_BOARD in your board file and rebuild. Most
likely on PPC at least it will hang, but if you are lucky it will print
something first :-) I hope to test an SPL board (snow / exynos5250) when
I can get that booting from mainline.

I have run this though buildman with CONFIG_SYS_GENERIC_BOARD on for all
ARM, PPC and x86 boards. The only failure is highbank (an ARM board), which
seems to use SCSI but scsi_init() is not available.

Code size increases by about 1KB on ARM and about 1.3KB on PPC with generic
board enabled. This is mostly due to the move to using separate functions
for each part of the init, which will make it easier to move to a pure
initcall approach later if we want to.

Since this series was first sent there have been additions to the board
code for ARM and PPC. I have added in these additions. This is a manual
process, and I expect that people will find problems (and send patches)
as they try out their boards.
END

Change-Id: Ib36d5f90c1fb9396d9e34404e332abdf313de2a1