AHCI: Make sending the SETFEATURES_XFER command optional
This command doesn't really do anything when talking to a SATA device, and
sending it confuses some of them. This change makes sending the command
optional, and defaults to not. The situations where it should be sent are not
the common case.

With the standard SSD in the machine, here are some times with the option
turned off:
1. 8277
2. 8273
3. 8050

And turned on:
1. 8303
2. 8155
3. 8276

Sending that command seems to have no meaningful effect on performance.

BUG=chrome-os-partner:7714
TEST=Booted off an SSD that hadn't worked with U-Boot previously. This SSD
was lent to us by Grant, and has chips labelled as Toshiba NV6424,
Taiwan 11159AE P, and TC58NVG5D2FTA10.

(cherry picked from commit 65df619cdea0e1b589254072d562b3d0fdeaf3b5)

Change-Id: I1673335f5e9dd1b8768e22db3fa5a089c3701107
Reviewed-on: https://gerrit.chromium.org/gerrit/14934
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>