arm: Add pivot_to_board_init_r() to each start.S
We don't want this in a common file, or at least not yet, so add
this function to every start.S individually. The existing code tacked
on the end of a long relocation function and does not suit our needs
since it doesn't allow the address of board_init_r() to be passed in
and cannot be called from C since it expects values in registers r4
and above.

Series-changes: 4
- Split out start_call_board_init_r() addition into new patch
- Put start_call_board_init_r() into each start.S, sadly

Series-changes: 5
- Rename start_call_board_init_r() to pivot_to_board_init_r()

Change-Id: I7cfcb5c33fde35567a9d31689c8de51b669a6b8e