arm: Add explicit __image_copy_end symbol for ARM926EJ-S
The start of the BSS is normally the end of the region in the image
that must be copied for relocation to work. But for SPL this is apparently
not always true, and ARMv7 has an explicit __image_copy_end symbol for
this. Add this new symbol to ARM926EJ-S to support SPL under generic
relocation.

Change-Id: I70b26dac252bdf7594b2dc993967ac4872346fc2