arm: Don't use printf() in SPL builds
raise() likes to call printf() if it is available, but in SPL builds it
either is not available, or adds a large chunk to the resulting image
size.

So don't call it even if it is available.

This change reduces SPL size from 10KB to 6.3KB on hawkboard, for
example, using generic relocation.

Series-to: u-boot
Series-cc: albert, hs
Series-cc: Sughosh Ganu <urwithsughosh@gmail.com>
Series-cc: Christian Riesch <christian.riesch@omicron.at>
Series-cc: tomr

Change-Id: Ieb3063b0bfdc3469b2d9155aeebf381e76859703