Cherry-Pick: gen: Make the SCSI and USB drivers compile with the CONFIG_SYS_64BIT_LBA option
These drivers wouldn't compile when the CONFIG_SYS_64BIT_LBA option is turned
on because the used fixed size data types in their exported functions when they
should have used lbaint_t for the block count parameter. That meant that when
the sizes happened to be the same, when using a 28 bit LBA, the driver would
build, but when it wasn't, a 48 bit LBA, things broke.

This change adjusts the signatures to use the right type and makes small
adjustments in the affected functions.

BUG=chrome-os-partner:8904
TEST=Built for emeraldlake2 with the CONFIG_SYS_64BIT_LBA option enabled and
saw its compile errors go away.

Change-Id: I9bcdcab52a34a1542febf2bfa79226888475fdc6
Original-Change-Id: I7443b5976d3d06b82070344011fba46a28bd77de
Signed-off-by: Gabe Black <gabeblack@google.com>
Original-Reviewed-on: https://gerrit.chromium.org/gerrit/18058
Original-Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/20968
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>