Cherry-Pick: gen: Get disk/part_dos.c to compile when CONFIG_SYS_64BIT_LBA is turned on
The code in this file tries to print an LBA using the format specifier
"%10ld", but when CONFIG_SYS_64BIT_LBA is turned on, the argument passed in
this position is a 64 bit value and isn't compatible. This change selects the
appropriate format specifier, the original or "10lld", based on whether
CONFIG_SYS_64BIT_LBA is turned on.

BUG=chrome-os-partner:8904
TEST=Built for Stumpy with BUILD_PART_FS_STUFF turned on and
CONFIG_SYS_64BIT_LBA enabled and disable, and saw the compile errors go away.

Change-Id: I4254a2a13d9b03f009d8e68cb0d1868f4033c578
Original-Change-Id: Ib80b1567ce07673aea2b47dfba9b3526b32f9de5
Signed-off-by: Gabe Black <gabeblack@google.com>
Original-Reviewed-on: https://gerrit.chromium.org/gerrit/18219
Original-Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20972
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>