temp: exynos: Add timer_get_us function
timer_get_us returns the time in microseconds since a certain reference
point of history.  However, it does not guarantee to return an accurate
time after a long period; instead, it wraps around (that is, the
reference point is reset to some other point of history) after some
periods.  The frequency of wrapping around is about an hour (or 2^32
microseconds).

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chromium-os:28077
TEST=Build okay

Change-Id: I8565d6726be784ba2b38fe13d3e2a3ec701befb3
Reviewed-on: https://gerrit.chromium.org/gerrit/20431
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>