temp: exynos: Collect and display board revision information
Add this feature both in SPL and in normal U-Boot. We read the board
revision information and display it in U-Boot. While we don't display
it in SPL, it is available when needed.

BUG=chrome-os-partner:9653
TEST=manual:
build and boot on daisy, see that it is 4 in both SPL, normal
Tested on modded snow, see that it is 6 in both SPL, normal
Boot kernel to login prompt on daisy (display does not work)

Change-Id: I5f347d0c691bce8a1da556c7652791338db2e29a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22625