(firmware-Parrot)coreboot/hda: Get beep node from codec.
Instead of hard-coding the beep command, find the nid by interrogating
the codec.  Find the beep node in the codec if it exists and send the
command to that node.  This allows for Link to beep without breaking
Stumpy/Lumpy.

BUG=chrome-os-partner:9037
TEST=hard-code Beep in vboot and test that it beeps for stumpy and link.

original-commit-id: c35e1e5fd20098d94a3d49a76f62739cd1901b26
Fixed a few things that git hook had issues with.

Change-Id: Ibaab99185f8cc04059564a7e7d3427b26c56d00d
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27248
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Signed-off-by: Marc Jones <marc.jones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30975
Commit-Ready: Jay Kim <yongjaek@chromium.org>
Reviewed-by: Jay Kim <yongjaek@chromium.org>
Tested-by: Jay Kim <yongjaek@chromium.org>