drop: exynos: fdt: Use elpida memory settings for all boards
The settings labelled "elpida" are not actually just for elpida
memory.  The differences between the elpida and samsung tables
appear to have more to do with board layout and memory packaging
than anything else.

We need to get to the bottom of this, but for now the suggestion is to use
the elpida settings for all Snow boards, even the samsung ones.

BUG=chrome-os-partner:9864
TEST=Can boot samsung board to prompt.

Change-Id: I4866e1067af995bcb664be519d34cf95acbe2d7d
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23239
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>
(cherry picked from commit 618c71dbcab46cb705b256b5c6ffa7841544dfab)
Reviewed-on: https://gerrit.chromium.org/gerrit/23600