Revert "CHERRY-PICK: drop: exynos: tmu: Hack on hardware thermal tripping."
This reverts commit 735e8a2156d4b3777b47e7ed04329b7e6304dfd4

It's believed that commit was interfering with the software
thermal trip code when a warm reset happens.  Specifically
I think the software thermal trip code in 
drivers/power/exynos-tmu.c effectively clears THRESH_LEVEL_RISE3, which may be a problem when the 
hardware thermal trip is set).  That needs to be investigated,
but for now the revert seems safest.

Change-Id: I44c9d62f64e1d76d2d66c7a3f6beb6ec66407578
Reviewed-on: https://gerrit.chromium.org/gerrit/25977
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>