CHERRY-PICK: cros: Clear keyboard buffer before starting to boot
We are seeing an issue where a transient keypress is apparently
reported by the EC on boot. The root cause of this needs to be
investigated but apparently it happens more than 10% of the time.

The EC should not deliver spurious characters on reset, and this
needs to be investigated. It may simply be that we need to reset the
EC's keyboard fifo when powering on the AP. But I have seen that
the EC fifo is generally full when booting into Linux, which seems
odd. It may be a bug related to the EC fifo.

This workaround may correct the problem, but needs testing. It clears
the keyboard buffer when starting twostop boot (both at the first
start and in the secondary U-Boot).

The call to tstc() ends up calling kbd_tstc() in mkbp_key.c, which
will check for an EC interrupt, and if there is one, it will add the
key to the U-Boot input fifo via input_tstc() which calls
check_for_keys(). The subsequent getc() will then read the key.

The interrupt line will remain active while there are keys to read,
as we have moved to level interrupts in the EC.

So simply calling tstc() again should bring out the next key if there
is any. In this way the EC keyboard fifo should be drained before we
look for keys which might affect secure boot.

BUG=chrome-os-partner:11075
TEST=manual
Please perform these steps at least 20-30 times:
Repeatedly power on the system with power button, and boot the
system. See that it does not go into recovery.

Change-Id: I8ec34d359e275a6c30c2d3bdba5053b3950ee8f1
Original-Change-Id: I31cadadf741702b8748b7ee6bef785fb959fb95c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27379
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Tested-by: Katie Roberts-Hoffman <katierh@chromium.org>