gen: mkbp: Refactor ec_command() to avoid copy
Now that the various transports can support copy-free operations,
extend this to callers of ec_command() as well.

The only interface change here is to use uint8_t **dinp for the response.
If this is NULL, then *dinp will be set to point to the response
parameters (64-bit aligned). If this is non-NULL, the response
parameters will be copied to *dinp.

BUG=chrome-os-partner:11319
TEST='mkbp test' in U-Boot, see that it succeeds

Original-Change-Id: I8a02edac7935440c5c32140a893bdc911b2e84d6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27486
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit 0893238e1c93686e12b18c1fa932a9e2debbcad0)

Change-Id: I30c69d4388c1d2c9b9e95b5c3047f6b67718eb69
Reviewed-on: https://gerrit.chromium.org/gerrit/29785
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>