gen: mkbp: Refactor lpc part to avoid copying
Adjust the lpc comms function to return a pointer to its input data,
instead of copying it.

Revert to using the required values from the fdt, and try to use the
common mkbp functions for comms trace and and checksumming.

The outsb() and insb() functions in U-Boot don't appear to work, so
we need to do our own loop to send/receive data over LPC. Put this
code into a separate function which we will soon use for parameters.

There are rather a lot of changes here, as this pre-dates the recent
mkbp refactor.

BUG=chrome-os-partner:11319
TEST=manual:
'mkbp test' in U-Boot on link, see that is succeeds
Build and boot on link, see that 'mkbp info' still works
Build with DEBUG enabled in driver, see that the message trace looks
reasonable:
cmd: 02, ver: 00
6c 69 6e 6b 5f 76 31 2e 31 2e 33 30 34 2d 33 63 63 34 37 34 35 00 42 71 04 02 59 aa bf 4a cb b5 6c 69 6e 6b 5f 76 31 2e 31 2e 33 30 34 2d 33 63 63 34 37 34 35 00 a4 32 aa ab b0 88 d1 ad e9 41 10 d1 b2 1f b4 c3 fe 23 98 13 5e 12 3a 51 27 f2 09 6f 43 15 28 85 c0 2d e9 ec 37 8a bf 67 e3 d8 01 00 00 00
cmd: 15, ver: 0144 00 00 00 44 00 00 00
08 00 00 00 3f 00 00 00 05 00 00 00
cmd: d2, ver: 0005 00

Original-Change-Id: Ibdd73ecfd2d7c1d331bff82f53e75b6c950ee22e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27484
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit 6dca26b587805ba902dbced20351dd903d754959)

Change-Id: I7ee142028201baca638e857554b54d7d13dcdd76
Reviewed-on: https://gerrit.chromium.org/gerrit/29784
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>