exynos: Move CPU detection code out of header file
This doesn't really need to be inline. There is only one caller
(in gpio.h). Move it to a C file.

Also there is really no need to have this in the lowlevel init, so
move it to arch_cpu_init(), which we will enable in a later commit.

BUG=chrome-os-partner:12153
TEST=manual
build and boot to kernel on snow

Original-Change-Id: Ib2200a5815b7ba80b4b282bf138bc699afa59118
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29539
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
(cherry picked from commit 99dcb84173d59a4eb112c55c9d2f14130ab05bf5)

Change-Id: Idd82ac4073a9b58747c044c6ca6f541a8804af39
Reviewed-on: https://gerrit.chromium.org/gerrit/29786
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>