exynos: fdt: add bad-wake-gpios to config
This GPIO is used to detect when a wakeup is invalid.  An invalid
wakeup happens when the state of this GPIO is high during wakeup.
This is used to detect a malicious user who might be trying to mess
with the "is sleeping" GPIO.  We detect this in U-Boot upon wakeup and
reboot the system if we find this attack.

BUG=chrome-os-partner:11639
TEST=Compiled.  Nothing is using this FDT entry at the moment
so no further tests are possible.  Eventually when cros_bundle_firmware
is updated we'll be able to re-run the tests from
<https://gerrit.chromium.org/gerrit/#/c/28612/> and make sure that they
still work (AKA that cros_bundle_firmware doesn't clobber things).

Original-Change-Id: Ie1e7f43049157703f9f6d0f161fa2d011306a11b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28946
Reviewed-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit 8836bbfa0aea80c980731000a3a23fb9408ea5d8)

Change-Id: Idb70649441b9bee01f833b52ff7ecb1c85ab0ea6
Reviewed-on: https://gerrit.chromium.org/gerrit/29515
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>