exynos: serial: Make the serial_s5p driver recognize the silent_console option
We need to get silent console working in U-Boot. That's important for a
number of reasons, but among them is that this is the way that we silence
the Linux console (see fixup_silent_linux).

BUG=chrome-os-partner:12074
TEST=Built and booted with silent_console on and off on snow. When off,
console output appeared as normal. With it on, console output was
suppressed in U-Boot.

Original-Change-Id: I00e70f8a24a84020a916b9142e25d9f82b62ee49
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/29007
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
(cherry picked from commit 378872657e13c2f5cfa3ba516b6fe8860d4dccf5)

Change-Id: I7c1b914c63b874b714620ec54c9ae2dd52c3cc33
Reviewed-on: https://gerrit.chromium.org/gerrit/29518
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>