cros: Enable software sync for daisy
This ties the EC firmware to the BIOS. When the RW BIOS is updated, the next
boot will force the EC firwmare to be updated to match.

In addition, because communication with daisy's EC is slow, the BIOS will
display a "please wait" screen while this is happening.

BUG=chrome-os-partner:10720
TEST=manual

First you have to build a BIOS that uses the RW A/B, not just RO.

  cros_workon --board daisy start chromeos-u-boot chromeos-ec
  repo sync

Edit cros/dts/flashmap-4mb-rw.dtsi (in two places) to say

   preamble-flags = <0>;

Then build a new EC firmware and BIOS:

  emerge-daisy chromeos-u-boot chromeos-ec chromeos-bootimage

Install that BIOS onto daisy and reboot. The EC should be automatically
updated to match, and you should see the "Please wait" screen while it
happens.

Original-Change-Id: I73605b29aeb6a9446cf4a6a2676e8144027e0098
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29486
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
(cherry picked from commit e85a235a72a5423e5432df490c73ef498604db52)

Change-Id: I24226ca42275d5e967aeaeb8e29222684272676f
Reviewed-on: https://gerrit.chromium.org/gerrit/29783
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>