exynos: Refactor low_level code to clarify paths
The current code repeats itself because it needs to do clock init
for a wake from sleep but not a week from idle.

Refactor the code to use a action mask, to avoid the duplication.
None of this init is required outside SPL, so make it conditional on
that (the next commit moves the whole file to SPL-only).

BUG=chrome-os-partner:12153
TEST=manual:
build and boot to kernel on snow
Check that suspend/resume works correctly
TODO: need to test the idle resume

Original-Change-Id: I84d9db524cda4b1f3f14f8660ce7d77a58353c7b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29545
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
(cherry picked from commit 133f81af50658cf3b639cb58865497272c34e992)

Change-Id: I1842a7f6eedc3538c95c4b6fd95b626875768d3c
Reviewed-on: https://gerrit.chromium.org/gerrit/30155
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>