cros: connect 'vboot_flag_id' values and string representations
This change combines the definition of the enumeration 'vboot_flag_id'
and the textual representation of these identifiers.

This makes the code less error prone since the defintion of the
enumerate identifiers and the initialized array containing the strings
will always be properly initialized.  It also makes it easier to see
the relationship between the identifier and the string when they are
in the same place.

BUG=chromium-os:32603
TEST=u-boot still builds; no code changes, no data changes.
Signed-off-by: Taylor Hutt <thutt@chromium.org>
Original-Change-Id: I2faa9065f176fdcfda9069632a3ee3ffa0c51d27
Reviewed-on: https://gerrit.chromium.org/gerrit/29814
Reviewed-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit 31f61d8dc5da366b870a307d639e5339a83b3b7c)

Change-Id: I7194e88ef6b49e72bb3bef04df8c2367287277b3
Reviewed-on: https://gerrit.chromium.org/gerrit/29937
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>